requirements.txt
ollama
python-dotenvpip 